add_theme {ggplot2}	R Documentation

Modify properties of an element in a theme object

Description

Modify properties of an element in a theme object

Usage

add_theme(t1, t2, t2name, call = caller_env())

Arguments

t1
A theme object

t2
A theme object that is to be added to t1

t2name
A name of the t2 object. This is used for printing informative error messages.

[Package ggplot2 version 3.4.4 Index]