aes_string {ggplot2}	R Documentation

Generate aesthetic mappings from a string/quoted objects

Description

Aesthetic mappings describe how variables in the data are mapped to visual properties (aesthetics) of geoms. aes uses non-standard evaluation to capture the variable names. These two variants use regular evaluation, which is easier to use inside functions.

Usage

aes_string(x = NULL, y = NULL, ...)

aes_q(x = NULL, y = NULL, ...)

Arguments

x,y,...
List of name value pairs

Details

aes_string and aes_q are particularly useful when writing functions that create plots because you can use strings or quoted names/calls to define the aesthetic mappings, rather than having to use substitute to generate a call to aes().

See Also

aes

Other aesthetic generators: aes

Examples

# Threee ways of generating the same aesthetics
aes(mpg, wt, col = cyl, fill = NULL)
aes_string("mpg", "wt", col = "cyl", fill = NULL)
aes_q(quote(mpg), quote(wt), col = quote(cyl), fill = NULL)

aes(col = cyl, fill = NULL)
aes_string(col = "cyl", fill = NULL)
aes_q(col = quote(cyl), fill = NULL)

[Package ggplot2 version 1.0.1 Index]