as.list.ggproto {ggplot2}	R Documentation

Convert a ggproto object to a list

Description

This will not include the object's super member.

Usage

## S3 method for class 'ggproto'
as.list(x, inherit = TRUE, ...)

Arguments

x
A ggproto object to convert to a list.

inherit
If TRUE (the default), flatten all inherited items into the returned list. If FALSE, do not include any inherited items.

...
Arguments passed on to base::as.list.environment

all.names

a logical indicating whether to copy all values or (default) only those whose names do not begin with a dot.

sorted

a logical indicating whether the names of the resulting list should be sorted (increasingly). Note that this is somewhat costly, but may be useful for comparison of environments.

[Package ggplot2 version 3.4.4 Index]