autoplot {ggplot2}	R Documentation

Create a complete ggplot appropriate to a particular data type

Description

autoplot() uses ggplot2 to draw a particular plot for an object of a particular class in a single command. This defines the S3 generic that other classes and packages can extend.

Usage

autoplot(object, ...)

Arguments

object
an object, whose class will determine the behaviour of autoplot

...
other arguments passed to specific methods

Value

a ggplot object

See Also

autolayer(), ggplot() and fortify()

[Package ggplot2 version 3.3.4 Index]