class_ggplot_built {ggplot2}	R Documentation

The ggplot built class

Description

The ggplot built class is an intermediate class and represents a processed ggplot object ready for rendering. It is constructed by calling ggplot_build() on a ggplot object and is not meant to be instantiated directly. The class can be rendered to a gtable object by calling the ggplot_gtable() function on a ggplot built class object.

Usage

class_ggplot_built(..., data = NULL, layout = NULL, plot = NULL)

Arguments

...
Reserved for future expansion.

data
A list of plain data frames; one for each layer.

layout
A Layout ggproto object.

plot
A completed ggplot class object.

[Package ggplot2 version 4.0.2 Index]