class_labels {ggplot2}	R Documentation

The labels class

Description

The labels class holds a list with label information to display as titles of plot components. The preferred way to construct an object of the labels class is to use the labs() function.

Usage

class_labels(labels = list(), ...)

Arguments

labels
A named list.

...
Reserved for future expansion.

Details

All members of labels are expected to be named and names should be unique.

[Package ggplot2 version 4.0.2 Index]