class_theme {ggplot2}	R Documentation

The theme class

Description

The theme class holds information on how non-data elements of the plot should be rendered. The preferred way to construct an object of this class is through the theme() function.

Usage

class_theme(elements = list(), ..., complete = FALSE, validate = TRUE)

Arguments

elements
A named list containing theme elements.

...
Reserved for future expansion.

complete
A boolean value stating whether a theme is complete.

validate
A boolean value stating whether a theme should still be validated.

[Package ggplot2 version 4.0.2 Index]