continuous_scale {ggplot2}	R Documentation

Continuous scale constructor.

Description

Continuous scale constructor.

Usage

  continuous_scale(aesthetics, scale_name, palette,
    name = NULL, breaks = waiver(),
    minor_breaks = waiver(), labels = waiver(),
    legend = NULL, limits = NULL, rescaler = rescale,
    oob = censor, expand = waiver(), na.value = NA_real_,
    trans = "identity", guide = "legend")

Arguments

minor_breaks
Used with date or datetime scales. Either NULL for no minor breaks, waiver() for the default breaks (one minor break between each major break), a numeric vector of positions, or a function that given the limits returns a vector of minor breaks.

limits
A numeric vector of length two describing the scale limits.

rescaler
Used by diverging and n colour gradients (i.e. scale_colour_gradient2, scale_colour_gradientn).

oob
What to do with values outside scale limits (out of bounds)?

aesthetics
the names of the aesthetics that this scale works with

scale_name
the name of the scale

palette
a palette function that when called with a single integer argument (the number of levels in the scale) returns the values that they should take

name
the name of the scale - used as the axis label or the legend title

breaks
control the breaks in the guide. There are four possible types of input:

NULL: don't display any breaks

a character vector giving the breaks as they should appear on the axis or in the legend.

waiver() to use the default break computation.

a function, that when called with a single argument, a character vector giving the limits of the scale, returns a character vector specifying which breaks to display.

This parameter does not affect in any way how the data is scaled - it only affects the appearance of the legend.

labels
NULL for no labels, waiver() for default labels (labels the same as breaks), a character vector the same length as breaks, or a named character vector whose names are used to match replacement the labels for matching breaks.

legend
deprecated. Use guide instead.

expand
a numeric vector of length two, giving a multiplicative and additive constant used to expand the range of the scales so that there is a small gap between the data and the axes.

na.value
how should missing values be displayed?

guide
the name of, or actual function, used to create the guide.

[Package ggplot2 version 0.9.3.1 Index]