coord_expand_defaults {ggplot2}	R Documentation

Set the default expand values for the scale, if NA

Description

Set the default expand values for the scale, if NA

Usage

  coord_expand_defaults(coord, scale, aesthetic = NULL)

[Package ggplot2 version 0.9.3.1 Index]