cut_number {ggplot2}	R Documentation

Cut numeric vector into intervals containing equal number of points.

Description

Cut numeric vector into intervals containing equal number of points.

Usage

  cut_number(x, n = NULL, ...)

Arguments

x
numeric vector

n
number of intervals to create

...
other arguments passed on to cut

See Also

cut_interval

Examples

table(cut_number(runif(1000), n = 10))

[Package ggplot2 version 0.9.3.1 Index]