datetime_scale {ggplot2}	R Documentation

Date/time scale constructor

Description

Date/time scale constructor

Usage

datetime_scale(
  aesthetics,
  trans,
  palette,
  breaks = pretty_breaks(),
  minor_breaks = waiver(),
  labels = waiver(),
  date_breaks = waiver(),
  date_labels = waiver(),
  date_minor_breaks = waiver(),
  timezone = NULL,
  guide = "legend",
  ...
)

Arguments

aesthetics
The names of the aesthetics that this scale works with.

trans
For date/time scales, the name of a date/time transformation or the object itself. Built-in transformations include "hms", "date" and "time".

palette
A palette function that when called with a numeric vector with values between 0 and 1 returns the corresponding output values (e.g., scales::area_pal()).

breaks
One of:

NULL for no breaks

waiver() for the breaks specified by date_breaks

A Date/POSIXct vector giving positions of breaks

A function that takes the limits as input and returns breaks as output

minor_breaks
One of:

NULL for no breaks

waiver() for the breaks specified by date_minor_breaks

A Date/POSIXct vector giving positions of minor breaks

A function that takes the limits as input and returns minor breaks as output

labels
One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

An expression vector (must be the same length as breaks). See ?plotmath for details.

A function that takes the breaks as input and returns labels as output. Also accepts rlang lambda function notation.

date_breaks
A string giving the distance between breaks like "2 weeks", or "10 years". If both breaks and date_breaks are specified, date_breaks wins. Valid specifications are 'sec', 'min', 'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.

date_labels
A string giving the formatting specification for the labels. Codes are defined in strftime(). If both labels and date_labels are specified, date_labels wins.

date_minor_breaks
A string giving the distance between minor breaks like "2 weeks", or "10 years". If both minor_breaks and date_minor_breaks are specified, date_minor_breaks wins. Valid specifications are 'sec', 'min', 'hour', 'day', 'week', 'month' or 'year', optionally followed by 's'.

timezone
The timezone to use for display on the axes. The default (NULL) uses the timezone encoded in the data.

guide
A function used to create a guide or its name. See guides() for more information.

...
Arguments passed on to continuous_scale

scale_name

The name of the scale that should be used for error messages associated with this scale.

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

n.breaks

An integer guiding the number of major breaks. The algorithm may choose a slightly different number to ensure nice break labels. Will only have an effect if breaks = waiver(). Use NULL to use the default number of breaks given by the transformation.

limits

One of:

NULL to use the default scale range

A numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum

A function that accepts the existing (automatic) limits and returns new limits. Also accepts rlang lambda function notation. Note that setting limits on positional scales will remove data outside of the limits. If the purpose is to zoom, use the limit argument in the coordinate system (see coord_cartesian()).

rescaler

A function used to scale the input values to the range [0, 1]. This is always scales::rescale(), except for diverging and n colour gradients (i.e., scale_colour_gradient2(), scale_colour_gradientn()). The rescaler is ignored by position scales, which always use scales::rescale(). Also accepts rlang lambda function notation.

oob

One of:

Function that handles limits outside of the scale limits (out of bounds). Also accepts rlang lambda function notation.

The default (scales::censor()) replaces out of bounds values with NA.

scales::squish() for squishing out of bounds values into range.

scales::squish_infinite() for squishing infinite values into range.

na.value

Missing values will be replaced with this value.

expand

For position scales, a vector of range expansion constants used to add some padding around the data to ensure that they are placed some distance away from the axes. Use the convenience function expansion() to generate the values for the expand argument. The defaults are to expand the scale by 5% on each side for continuous variables, and by 0.6 units on each side for discrete variables.

position

For position scales, The position of the axis. left or right for y axes, top or bottom for x axes.

super

The super class to use for the constructed scale

[Package ggplot2 version 3.4.4 Index]