element_blank {ggplot2}	R Documentation

Theme element: blank. This theme element draws nothing, and assigns no space

Description

Theme element: blank. This theme element draws nothing, and assigns no space

Usage

  element_blank()

[Package ggplot2 version 0.9.3.1 Index]