element_render {ggplot2}	R Documentation

Render a specified theme element into a grob

Description

Given a theme object and element name, returns a grob for the element. Uses element_grob() to generate the grob.

Usage

element_render(theme, element, ..., name = NULL)

Arguments

theme
The theme object

element
The element name given as character vector

...
Other arguments provided to element_grob()

name
Character vector added to the name of the grob

[Package ggplot2 version 3.3.5 Index]