element_text {ggplot2}	R Documentation

Theme element: text.

Description

Theme element: text.

Usage

element_text(family = NULL, face = NULL, colour = NULL, size = NULL,
  hjust = NULL, vjust = NULL, angle = NULL, lineheight = NULL,
  color = NULL, margin = NULL, debug = NULL)

Arguments

family
font family

face
font face ("plain", "italic", "bold", "bold.italic")

colour
text colour

size
text size (in pts)

hjust
horizontal justification (in [0, 1])

vjust
vertical justification (in [0, 1])

angle
angle (in [0, 360])

lineheight
line height

color
an alias for colour

margin
margins around the text. See margin for more details. When creating a theme, the margins should be placed on the side of the text facing towards the center of the plot.

debug
If TRUE, aids visual debugging by drawing a solid rectangle behind the complete text area, and a point where each label is anchored.

[Package ggplot2 version 2.1.0 Index]