fill_alpha {ggplot2}	R Documentation

Modify fill transparency

Description

This works much like alpha() in that it modifies the transparency of fill colours. It differs in that fill_alpha() also attempts to set the transparency of <GridPattern> objects.

Usage

fill_alpha(fill, alpha)

Arguments

fill
A fill colour given as a character or integer vector, or as a (list of) <GridPattern> object(s).

alpha
A transparency value between 0 (transparent) and 1 (opaque), parallel to fill.

Value

A character vector of colours, or list of <GridPattern> objects.

Examples

# Typical colour input
fill_alpha("red", 0.5)

if (utils::packageVersion("grid") > "4.2") {
  # Pattern input
  fill_alpha(list(grid::linearGradient()), 0.5)
}

[Package ggplot2 version 3.5.1 Index]