fortify {ggplot2}	R Documentation

Fortify a model with data.

Description

Method to convert a generic R object into a data frame useful for plotting. Takes its name from the idea of fortifying the original data with model fit statistics, and vice versa.

Usage

  fortify(model, data, ...)

Arguments

model
model or other R object to convert to data frame

data
original dataset, if needed

...
other arguments passed to methods

See Also

fortify.lm

[Package ggplot2 version 0.9.3.1 Index]