geom_bin2d {ggplot2}	R Documentation

Add heatmap of 2d bin counts.

Description

Add heatmap of 2d bin counts.

Usage

  geom_bin2d(mapping = NULL, data = NULL, stat = "bin2d",
    position = "identity", ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_bin2d understands the following aesthetics (required aesthetics are in bold):

xmax

xmin

ymax

ymin

alpha

colour

fill

linetype

size

weight

Examples

d <- ggplot(diamonds, aes(x = x, y = y)) + xlim(4,10) + ylim(4,10)
d + geom_bin2d()
d + geom_bin2d(binwidth = c(0.1, 0.1))

# See ?stat_bin2d for more examples

[Package ggplot2 version 0.9.3.1 Index]