geom_density {ggplot2}	R Documentation

Display a smooth density estimate.

Description

A smooth density estimate calculated by stat_density.

Usage

  geom_density(mapping = NULL, data = NULL,
    stat = "density", position = "identity", na.rm = FALSE,
    ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

na.rm
If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_density understands the following aesthetics (required aesthetics are in bold):

x

y

alpha

colour

fill

linetype

size

weight

See Also

geom_histogram for the histogram and stat_density for examples.

Examples

# See stat_density for examples

[Package ggplot2 version 0.9.3.1 Index]