geom_errorbarh {ggplot2}	R Documentation

Horizontal error bars

Description

Horizontal error bars

Usage

  geom_errorbarh(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_errorbarh understands the following aesthetics (required aesthetics are in bold):

x

xmax

xmin

y

alpha

colour

height

linetype

size

See Also

geom_errorbar: vertical error bars

Examples

df <- data.frame(
  trt = factor(c(1, 1, 2, 2)),
  resp = c(1, 5, 3, 4),
  group = factor(c(1, 2, 1, 2)),
  se = c(0.1, 0.3, 0.3, 0.2)
)

# Define the top and bottom of the errorbars

p <- ggplot(df, aes(resp, trt, colour = group))
p + geom_point() +
  geom_errorbarh(aes(xmax = resp + se, xmin = resp - se))
p + geom_point() +
  geom_errorbarh(aes(xmax = resp + se, xmin = resp - se, height = .2))

[Package ggplot2 version 0.9.3.1 Index]