geom_linerange {ggplot2}	R Documentation

An interval represented by a vertical line.

Description

An interval represented by a vertical line.

Usage

  geom_linerange(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_linerange understands the following aesthetics (required aesthetics are in bold):

x

ymax

ymin

alpha

colour

linetype

size

See Also

geom_errorbar: error bars; geom_pointrange: range indicated by straight line, with point in the middle; geom_crossbar: hollow bar with middle indicated by horizontal line; stat_summary: examples of these guys in use; geom_smooth: for continuous analog

Examples

# Generate data: means and standard errors of means for prices
# for each type of cut
dmod <- lm(price ~ cut, data=diamonds)
cuts <- data.frame(cut=unique(diamonds$cut), predict(dmod, data.frame(cut = unique(diamonds$cut)), se=TRUE)[c("fit","se.fit")])

qplot(cut, fit, data=cuts)
# With a bar chart, we are comparing lengths, so the y-axis is
# automatically extended to include 0
qplot(cut, fit, data=cuts, geom="bar")

# Display estimates and standard errors in various ways
se <- ggplot(cuts, aes(cut, fit,
  ymin = fit - se.fit, ymax=fit + se.fit, colour = cut))
se + geom_linerange()
se + geom_pointrange()
se + geom_errorbar(width = 0.5)
se + geom_crossbar(width = 0.5)

# Use coord_flip to flip the x and y axes
se + geom_linerange() + coord_flip()

[Package ggplot2 version 0.9.3.1 Index]