geom_pointrange {ggplot2}	R Documentation

An interval represented by a vertical line, with a point in the middle.

Description

An interval represented by a vertical line, with a point in the middle.

Usage

  geom_pointrange(mapping = NULL, data = NULL,
    stat = "identity", position = "identity", ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_pointrange understands the following aesthetics (required aesthetics are in bold):

x

y

ymax

ymin

alpha

colour

fill

linetype

shape

size

See Also

geom_errorbar for error bars, geom_linerange for range indicated by straight line, + examples, geom_crossbar for hollow bar with middle indicated by horizontal line, stat_summary for examples of these guys in use, geom_smooth for continuous analog"

Examples

# See geom_linerange for examples

[Package ggplot2 version 0.9.3.1 Index]