geom_quantile {ggplot2}	R Documentation

Add quantile lines from a quantile regression.

Description

This can be used as a continuous analogue of a geom_boxplot.

Usage

  geom_quantile(mapping = NULL, data = NULL,
    stat = "quantile", position = "identity",
    lineend = "butt", linejoin = "round", linemitre = 1,
    na.rm = FALSE, ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

na.rm
If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

lineend
Line end style (round, butt, square)

linejoin
Line join style (round, mitre, bevel)

linemitre
Line mitre limit (number greater than 1)

Aesthetics

geom_quantile understands the following aesthetics (required aesthetics are in bold):

x

y

alpha

colour

linetype

size

weight

See Also

See stat_quantile for examples.

Examples

# See stat_quantile for examples

[Package ggplot2 version 0.9.3.1 Index]