geom_rug {ggplot2}	R Documentation

Marginal rug plots.

Description

Marginal rug plots.

Usage

  geom_rug(mapping = NULL, data = NULL, stat = "identity",
    position = "identity", sides = "bl", ...)

Arguments

sides
A string that controls which sides of the plot the rugs appear on. It can be set to a string containing any of "trbl", for top, right, bottom, and left.

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

stat
The statistical transformation to use on the data for this layer.

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

geom_rug understands the following aesthetics (required aesthetics are in bold):

alpha

colour

linetype

size

Examples

p <- ggplot(mtcars, aes(x=wt, y=mpg))
p + geom_point()
p + geom_point() + geom_rug()
p + geom_point() + geom_rug(sides="b")    # Rug on bottom only
p + geom_point() + geom_rug(sides="trbl") # All four sides
p + geom_point() + geom_rug(position='jitter')

[Package ggplot2 version 0.9.3.1 Index]