get_guide_data {ggplot2}	R Documentation

Extract tick information from guides

Description

get_guide_data() builds a plot and extracts information from guide keys. This information typically contains positions, values and/or labels, depending on which aesthetic is queried or guide is used.

Usage

get_guide_data(plot = last_plot(), aesthetic, panel = 1L)

Arguments

plot
A ggplot or ggplot_build object.

aesthetic
A string that describes a single aesthetic for which to extract guide information. For example: "colour", "size", "x" or "y.sec".

panel
An integer giving a panel number for which to return position guide information.

Value

One of the following:

A data.frame representing the guide key, when the guide is unique for the aesthetic.

A list when the coord does not support position axes or multiple guides match the aesthetic.

NULL when no guide key could be found.

Examples

# A standard plot
p <- ggplot(mtcars) +
  aes(mpg, disp, colour = drat, size = drat) +
  geom_point() +
  facet_wrap(vars(cyl), scales = "free_x")

# Guide information for legends
get_guide_data(p, "size")

# Note that legend guides can be merged
merged <- p + guides(colour = "legend")
get_guide_data(merged, "size")

# Guide information for positions
get_guide_data(p, "x", panel = 2)

# Coord polar doesn't support proper guides, so we get a list
polar <- p + coord_polar()
get_guide_data(polar, "theta", panel = 2)

[Package ggplot2 version 3.5.1 Index]