gg_dep {ggplot2}	R Documentation

Give a deprecation error, warning, or message, depending on version number.

Description

Usage

gg_dep(version, msg)

Arguments

version
The last version of ggplot2 where this function was good (in other words, the last version where it was not deprecated).

msg
The message to print.

[Package ggplot2 version 3.4.4 Index]