ggorder {ggplot2}	R Documentation

A plot to investigate the order in which observations were recorded.

Description

A plot to investigate the order in which observations were recorded.

Usage

  ggorder(data)

Arguments

data
data set to plot

[Package ggplot2 version 0.9.3.1 Index]