ggplot.data.frame {ggplot2}	R Documentation

Create a new ggplot plot from a data frame

Description

Create a new ggplot plot from a data frame

Usage

  ## S3 method for class 'data.frame'
 ggplot(data, mapping = aes(), ...,
    environment = globalenv())

Arguments

data
default data frame for plot

mapping
default list of aesthetic mappings (these can be colour, size, shape, line type – see individual geom functions for more details)

...
ignored

environment
in which evaluation of aesthetics should occur

See Also

http://had.co.nz/ggplot2

[Package ggplot2 version 0.9.3.1 Index]