ggplot_build {ggplot2}	R Documentation

Build ggplot for rendering.

Description

This function takes the plot object, and performs all steps necessary to produce an object that can be rendered. This function outputs two pieces: a list of data frames (one for each layer), and a panel object, which contain all information about axis limits, breaks etc.

Usage

  ggplot_build(plot)

Arguments

plot
ggplot object

See Also

print.ggplot and link{benchplot} for for functions that contain the complete set of steps for generating a ggplot2 plot.

[Package ggplot2 version 0.9.3.1 Index]