ggstructure {ggplot2}	R Documentation

A plot which aims to reveal gross structural anomalies in the data.

Description

A plot which aims to reveal gross structural anomalies in the data.

Usage

  ggstructure(data)

Arguments

data
data set to plot

Examples

## Not run: 
ggstructure(mtcars)

## End(Not run)

[Package ggplot2 version 0.9.3.1 Index]