ggtheme {ggplot2}	R Documentation

ggplot2 themes

Description

Themes set the general aspect of the plot such as the colour of the background, gridlines, the size and colour of fonts.

Usage

theme_grey(base_size = 12, base_family = "")

theme_gray(base_size = 12, base_family = "")

theme_bw(base_size = 12, base_family = "")

theme_linedraw(base_size = 12, base_family = "")

theme_light(base_size = 12, base_family = "")

theme_minimal(base_size = 12, base_family = "")

theme_classic(base_size = 12, base_family = "")

Arguments

base_size
base font size

base_family
base font family

Details

theme_gray

The signature ggplot2 theme with a grey background and white gridlines, designed to put the data forward yet make comparisons easy.

theme_bw

The classic dark-on-light ggplot2 theme. May work better for presentations displayed with a projector.

theme_linedraw

A theme with only black lines of various widths on white backgrounds, reminiscent of a line drawings. Serves a purpose similar to theme_bw. Note that this theme has some very thin lines (<< 1 pt) which some journals may refuse.

theme_light

A theme similar to theme_linedraw but with light grey lines and axes, to direct more attention towards the data.

theme_minimal

A minimalistic theme with no background annotations.

theme_classic

A classic-looking theme, with x and y axis lines and no gridlines.

Examples

p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
     colour=factor(gear))) + facet_wrap(~am)

p
p + theme_gray()
p + theme_bw()
p + theme_linedraw()
p + theme_light()
p + theme_minimal()
p + theme_classic()

[Package ggplot2 version 1.0.1 Index]