guide-exts {ggplot2}	R Documentation

S3 generics for guides.

Description

You will need to provide methods for these S3 generics if you want to create your own guide object. They are currently undocumented; use at your own risk!

Usage

guide_train(guide, scale, aesthetic = NULL)

guide_merge(guide, new_guide)

guide_geom(guide, layers, default_mapping)

guide_transform(guide, coord, panel_params)

guide_gengrob(guide, theme)

Arguments

guide
The guide object

[Package ggplot2 version 3.4.4 Index]