hmisc {ggplot2}	R Documentation

Wrap up a selection of summary functions from Hmisc to make it easy to use with stat_summary.

Description

See the Hmisc documentation for details of their options.

Usage

  mean_cl_boot(x, ...)

  mean_cl_normal(x, ...)

  mean_sdl(x, ...)

  median_hilow(x, ...)

Arguments

x
a numeric vector

...
other arguments passed on to the respective Hmisc function.

See Also

smean.cl.boot, smean.cl.normal, smean.sdl, smedian.hilow

[Package ggplot2 version 0.9.3.1 Index]