last_plot {ggplot2}	R Documentation

Retrieve the last plot to be modified or created.

Description

Retrieve the last plot to be modified or created.

Usage

last_plot()

See Also

ggsave()

[Package ggplot2 version 3.3.4 Index]