margin {ggplot2}	R Documentation

Define margins.

Description

This is a convenient function that creates a grid unit object of the correct length to use for setting margins.

Usage

margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")

Arguments

t, r, b, l
Dimensions of each margin. (To remember order, think trouble).

unit
Default units of dimensions. Defaults to "pt" so it can be most easily scaled with the text.

Examples

margin(4)
margin(4, 2)
margin(4, 3, 2, 1)

[Package ggplot2 version 2.1.0 Index]