max_height {ggplot2}	R Documentation

Get the maximal width/length of a list of grobs

Description

Get the maximal width/length of a list of grobs

Usage

max_height(grobs, value_only = FALSE)

max_width(grobs, value_only = FALSE)

Arguments

grobs
A list of grobs

value_only
Should the return value be a simple numeric vector giving the maximum in cm

Value

The largest value. measured in cm as a unit object or a numeric vector depending on value_only

[Package ggplot2 version 3.4.4 Index]