merge_element {ggplot2}	R Documentation

Merge a parent element into a child element

Description

This is a generic and element classes must provide an implementation of this method

Usage

merge_element(new, old)

## Default S3 method:
merge_element(new, old)

## S3 method for class 'element_blank'
merge_element(new, old)

## S3 method for class 'element'
merge_element(new, old)

Arguments

new
The child element in the theme hierarchy

old
The parent element in the theme hierarchy

Value

A modified version of new updated with the properties of old

Examples

new <- element_text(colour = "red")
old <- element_text(colour = "blue", size = 10)

# Adopt size but ignore colour
merge_element(new, old)

[Package ggplot2 version 3.4.4 Index]