An updated and expanded version of the mammals sleep dataset.

This is an updated and expanded version of the mammals sleep dataset. Updated sleep times and weights were taken from V. M. Savage and G. B. West. A quantitative, theoretical framework for understanding mammalian sleep. Proceedings of the National Academy of Sciences, 104 (3):1051-1056, 2007.

msleep

Format

A data frame with 83 rows and 11 variables

name. common name

genus.

vore. carnivore, omnivore or herbivore?

order.

conservation. the conservation status of the animal

sleep\_total. total amount of sleep, in hours

sleep\_rem. rem sleep, in hours

sleep\_cycle. length of sleep cycle, in hours

awake. amount of time spent awake, in hours

brainwt. brain weight in kilograms

bodywt. body weight in kilograms

Details

Additional variables order, conservation status and vore were added from wikipedia.