plot-templates {ggplot2}	R Documentation

Plot templates.

Description

These plot templates are deprecated in an attempt to make ggplot2 as streamlined as possible, and to avoid bugs in these poorly tested functions. See the GGally package for some alternatives.

Usage

ggpcp(data, vars = names(data), ...)

ggfluctuation(table, type = "size", floor = 0, ceiling = max(table$freq,
  na.rm = TRUE))

ggmissing(data, avoid = "stack", order = TRUE, missing.only = TRUE)

ggstructure(data)

ggorder(data)

plotmatrix(data, mapping = aes(), colour = "black")

[Package ggplot2 version 1.0.1 Index]