position_fill {ggplot2}	R Documentation

Stack overlapping objects on top of one another, and standardise to have equal height.

Description

Stack overlapping objects on top of one another, and standardise to have equal height.

Usage

  position_fill(width = NULL, height = NULL)

Arguments

width
Manually specify width (does not affect all position adjustments)

height
Manually specify height (does not affect all position adjustments)

See Also

See geom_bar and geom_area for more examples.

Other position adjustments: position_dodge, position_identity, position_jitter, position_stack

Examples

# See ?geom_bar and ?geom_area for more examples
ggplot(mtcars, aes(x=factor(cyl), fill=factor(vs))) +
  geom_bar(position="fill")

cde <- geom_histogram(position="fill", binwidth = 500)

ggplot(diamonds, aes(x=price)) + cde
ggplot(diamonds, aes(x=price, fill=cut)) + cde
ggplot(diamonds, aes(x=price, fill=clarity)) + cde
ggplot(diamonds, aes(x=price, fill=color)) + cde

[Package ggplot2 version 0.9.3.1 Index]