presidential {ggplot2}	R Documentation

Terms of 11 presidents from Eisenhower to Obama.

Description

The names of each president, the start and end date of their term, and their party of 11 US presidents from Eisenhower to Obama.

Usage

presidential

Format

A data frame with 11 rows and 4 variables

[Package ggplot2 version 2.1.0 Index]