scale_area {ggplot2}	R Documentation

Scale area instead of radius (for size).

Description

scale_area is deprecated and will be removed in a future version of ggplot2. Use scale_size_area instead. Note that the default behavir of scale_size_area is slightly different: by default, it makes the area proportional to the numeric value.

Usage

  scale_area(..., range = c(1, 6))

Arguments

...
Other arguments passed on to continuous_scale to control name, limits, breaks, labels and so forth.

range
Range of output sizes. Should be greater than 0.

[Package ggplot2 version 0.9.3.1 Index]