scale_colour_grey {ggplot2}	R Documentation

Sequential grey colour scale.

Description

Based on gray.colors

Usage

  scale_colour_grey(..., start = 0.2, end = 0.8,
    na.value = "red")

  scale_fill_grey(..., start = 0.2, end = 0.8,
    na.value = "grey50")

  scale_color_grey(..., start = 0.2, end = 0.8,
    na.value = "red")

Arguments

start
gray value at low end of palette

end
gray value at high end of palette

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

na.value
Colour to use for missing values

See Also

Other colour scales: scale_color_brewer, scale_color_continuous, scale_color_discrete, scale_color_gradient, scale_color_gradient2, scale_color_gradientn, scale_color_hue, scale_colour_brewer, scale_colour_continuous, scale_colour_discrete, scale_colour_gradient, scale_colour_gradient2, scale_colour_gradientn, scale_colour_hue, scale_fill_brewer, scale_fill_continuous, scale_fill_discrete, scale_fill_gradient, scale_fill_gradient2, scale_fill_gradientn, scale_fill_hue

Examples

p <- qplot(mpg, wt, data=mtcars, colour=factor(cyl))
p + scale_colour_grey()
p + scale_colour_grey(end = 0)

# You may want to turn off the pale grey background with this scale
p + scale_colour_grey() + theme_bw()

# Colour of missing values is controlled with na.value:
miss <- factor(sample(c(NA, 1:5), nrow(mtcars), rep = TRUE))
qplot(mpg, wt, data = mtcars, colour = miss) + scale_colour_grey()
qplot(mpg, wt, data = mtcars, colour = miss) +
  scale_colour_grey(na.value = "green")

[Package ggplot2 version 0.9.3.1 Index]