scale_size_area {ggplot2}	R Documentation

Scale area instead of radius, for size.

Description

When scale_size_area is used, the default behavior is to scale the area of points to be proportional to the value.

Usage

  scale_size_area(..., max_size = 6)

Arguments

...
Other arguments passed on to continuous_scale to control name, limits, breaks, labels and so forth.

max_size
Size of largest points.

Details

Note that this controls the size scale, so it will also control the thickness of lines. Line thickness will be proportional to the square root of the value, which is probably undesirable in most cases.

[Package ggplot2 version 0.9.3.1 Index]