set_last_plot {ggplot2}	R Documentation

Set the last plot to be fetched by lastplot()

Description

Set the last plot to be fetched by lastplot()

Usage

set_last_plot(value)

See Also

last_plot()

[Package ggplot2 version 3.4.4 Index]