stat_ecdf {ggplot2}	R Documentation

Empirical Cumulative Density Function

Description

Empirical Cumulative Density Function

Usage

  stat_ecdf(mapping = NULL, data = NULL, geom = "step",
    position = "identity", n = NULL, ...)

Arguments

n
if NULL, do not interpolate. If not NULL, this is the number of points to interpolate with.

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

geom
The geometric object to use display the data

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Value

a data.frame with additional columns:

x
x in data

y
cumulative density corresponding x

Examples

qplot(rnorm(1000), stat = "ecdf", geom = "step")

df <- data.frame(x = c(rnorm(100, 0, 3), rnorm(100, 0, 10)),
                 g = gl(2, 100))

ggplot(df, aes(x, colour = g)) + stat_ecdf()

[Package ggplot2 version 0.9.3.1 Index]