stat_ellipse {ggplot2}	R Documentation

Plot data ellipses.

Description

Plot data ellipses.

Usage

stat_ellipse(mapping = NULL, data = NULL, geom = "path",
  position = "identity", type = "t", level = 0.95, segments = 51,
  na.rm = FALSE, ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

geom
The geometric object to use display the data

position
The position adjustment to use for overlappling points on this layer

type
The type of ellipse. The default "t" assumes a multivariate t-distribution, and "norm" assumes a multivariate normal distribution. "euclid" draws a circle with the radius equal to level, representing the euclidian distance from the center. This ellipse probably won't appear circular unless coord_fixed() is applied.

level
The confidence level at which to draw an ellipse (default is 0.95), or, if type="euclid", the radius of the circle to be drawn.

segments
The number of segments to be used in drawing the ellipse.

na.rm
If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Details

The method for calculating the ellipses has been modified from car::ellipse (Fox and Weisberg, 2011)

References

John Fox and Sanford Weisberg (2011). An R Companion to Applied Regression, Second Edition. Thousand Oaks CA: Sage. URL: http://socserv.socsci.mcmaster.ca/jfox/Books/Companion

Examples

ggplot(faithful, aes(waiting, eruptions))+
  geom_point()+
  stat_ellipse()

ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3))+
  geom_point()+
  stat_ellipse()

ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3))+
  geom_point()+
  stat_ellipse(type = "norm", linetype = 2)+
  stat_ellipse(type = "t")

ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3))+
  geom_point()+
  stat_ellipse(type = "norm", linetype = 2)+
  stat_ellipse(type = "euclid", level = 3)+
  coord_fixed()

ggplot(faithful, aes(waiting, eruptions, color = eruptions > 3))+
  stat_ellipse(geom = "polygon")

[Package ggplot2 version 1.0.1 Index]