stat_hline {ggplot2}	R Documentation

Add a horizontal line

Description

Add a horizontal line

Usage

  stat_hline(mapping = NULL, data = NULL, geom = "hline",
    position = "identity", yintercept, ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

geom
The geometric object to use display the data

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

See Also

geom_hline for code examples.

Examples

# see geom_hline

[Package ggplot2 version 0.9.3.1 Index]