stat_identity {ggplot2}	R Documentation

Identity statistic.

Description

Identity statistic.

Usage

  stat_identity(mapping = NULL, data = NULL,
    geom = "point", position = "identity", width = NULL,
    height = NULL, ...)

Arguments

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

geom
The geometric object to use display the data

position
The position adjustment to use for overlappling points on this layer

width
The width of the tiles.

height
The height of the tiles.

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Aesthetics

stat_identity understands the following aesthetics (required aesthetics are in bold):

Examples

# Doesn't do anything, so hard to come up a useful example

[Package ggplot2 version 0.9.3.1 Index]