stat_quantile {ggplot2}	R Documentation

Continuous quantiles.

Description

Continuous quantiles.

Usage

  stat_quantile(mapping = NULL, data = NULL,
    geom = "quantile", position = "identity",
    quantiles = c(0.25, 0.5, 0.75), formula = NULL,
    method = "rq", na.rm = FALSE, ...)

Arguments

quantiles
conditional quantiles of y to calculate and display

formula
formula relating y variables to x variables

method
Quantile regression method to use. Currently only supports rq.

na.rm
If FALSE (the default), removes missing values with a warning. If TRUE silently removes missing values.

mapping
The aesthetic mapping, usually constructed with aes or aes_string. Only needs to be set at the layer level if you are overriding the plot defaults.

data
A layer specific dataset - only needed if you want to override the plot defaults.

geom
The geometric object to use display the data

position
The position adjustment to use for overlappling points on this layer

...
other arguments passed on to layer. This can include aesthetics whose values you want to set, not map. See layer for more details.

Value

a data.frame with additional columns:

quantile
quantile of distribution

Aesthetics

stat_quantile understands the following aesthetics (required aesthetics are in bold):

x

y

Examples

msamp <- movies[sample(nrow(movies), 1000), ]
m <- ggplot(msamp, aes(year, rating)) + geom_point()
m + stat_quantile()
m + stat_quantile(quantiles = 0.5)
q10 <- seq(0.05, 0.95, by=0.05)
m + stat_quantile(quantiles = q10)

# You can also use rqss to fit smooth quantiles
m + stat_quantile(method = "rqss")
# Note that rqss doesn't pick a smoothing constant automatically, so
# you'll need to tweak lambda yourself
m + stat_quantile(method = "rqss", lambda = 10)
m + stat_quantile(method = "rqss", lambda = 100)

# Use 'votes' as weights for the quantile calculation
m + stat_quantile(aes(weight=votes))

# Change scale
m + stat_quantile(aes(colour = ..quantile..), quantiles = q10)
m + stat_quantile(aes(colour = ..quantile..), quantiles = q10) +
  scale_colour_gradient2(midpoint = 0.5)

# Set aesthetics to fixed value
m + stat_quantile(colour = "red", size = 2, linetype = 2)

# Use qplot instead
qplot(year, rating, data=movies, geom="quantile")

[Package ggplot2 version 0.9.3.1 Index]