theme_classic {ggplot2}	R Documentation

A classic-looking theme, with x and y axis lines and no gridlines.

Description

A classic-looking theme, with x and y axis lines and no gridlines.

Usage

  theme_classic(base_size = 12, base_family = "")

Arguments

base_size
base font size

base_family
base font family

[Package ggplot2 version 0.9.3.1 Index]