txhousing {ggplot2}	R Documentation

Housing sales in TX.

Description

Information about the housing market in Texas provided by the TAMU real estate center, http://recenter.tamu.edu/.

Usage

txhousing

Format

A data frame with 8602 observations and 9 variables:

cityName of MLS area

year,month,dateDate

salesNumber of sales

volumeTotal value of sales

medianMedian sale price

listingsTotal active listings

inventory"Months inventory": amount of time it would take to sell all current listings at current pace of sales.

[Package ggplot2 version 2.1.0 Index]