xlim {ggplot2}	R Documentation

Convenience functions to set the limits of the x and y axis.

Description

Observations not in this range will be dropped completely and not passed to any other layers.

Usage

  xlim(...)

  ylim(...)

Arguments

...
if numeric, will create a continuous scale, if factor or character, will create a discrete scale.

See Also

For changing x or y axis limits without dropping data observations, see coord_cartesian.

Examples

# xlim
xlim(15, 20)
xlim(20, 15)
xlim(c(10, 20))
xlim("a", "b", "c")
qplot(mpg, wt, data=mtcars) + xlim(15, 20)

# ylim
ylim(15, 20)
ylim(c(10, 20))
ylim("a", "b", "c")
qplot(mpg, wt, data=mtcars) + ylim(15, 20)

[Package ggplot2 version 0.9.3.1 Index]