zeroGrob {ggplot2}	R Documentation

The zero grob draws nothing and has zero size.

Description

The zero grob draws nothing and has zero size.

Usage

  zeroGrob()

[Package ggplot2 version 0.9.3.1 Index]