aes {ggplot2}	R Documentation

Generate aesthetic mappings that describe how variables in the data are mapped to visual properties (aesthetics) of geoms.

Description

aes creates a list of unevaluated expressions. This function also performs partial name matching, converts color to colour, and old style R names to ggplot names (eg. pch to shape, cex to size)

Usage

  aes(x, y, ...)

Arguments

x
x value

y
y value

...
List of name value pairs giving aesthetics to map.

See Also

aes_string for passing quoted variable names. aes_colour_fill_alpha, aes_group_order, aes_linetype_size_shape and aes_position for more specific examples with different aesthetics.

Examples

aes(x = mpg, y = wt)
aes(x = mpg ^ 2, y = wt / cyl)

[Package ggplot2 version 0.9.3.1 Index]