aes_all {ggplot2}	R Documentation

Given a character vector, create a set of identity mappings

Description

Given a character vector, create a set of identity mappings

Usage

aes_all(vars)

Arguments

vars
vector of variable names

Examples

aes_all(names(mtcars))
aes_all(c("x", "y", "col", "pch"))

[Package ggplot2 version 3.4.4 Index]