autolayer {ggplot2}	R Documentation

Create a ggplot layer appropriate to a particular data type

Description

autolayer() uses ggplot2 to draw a particular layer for an object of a particular class in a single command. This defines the S3 generic that other classes and packages can extend.

Usage

autolayer(object, ...)

Arguments

object
an object, whose class will determine the behaviour of autolayer

...
other arguments passed to specific methods

Value

a ggplot layer

See Also

autoplot(), ggplot() and fortify()

[Package ggplot2 version 3.3.5 Index]