benchplot {ggplot2}	R Documentation

Benchmark plot creation time. Broken down into construct, build, render and draw times.

Description

Benchmark plot creation time. Broken down into construct, build, render and draw times.

Usage

  benchplot(x)

Arguments

x
code to create ggplot2 plot

Examples

benchplot(qplot(mpg, wt, data = mtcars))
benchplot(qplot(mpg, wt, data = mtcars) + facet_grid(.~ cyl))

[Package ggplot2 version 0.9.3.1 Index]