calc_element {ggplot2}	R Documentation

Calculate the element properties, by inheriting properties from its parents

Description

Calculate the element properties, by inheriting properties from its parents

Usage

calc_element(element, theme, verbose = FALSE, skip_blank = FALSE)

Arguments

element
The name of the theme element to calculate

theme
A theme object (like theme_grey())

verbose
If TRUE, print out which elements this one inherits from

skip_blank
If TRUE, elements of type element_blank in the inheritance hierarchy will be ignored.

Examples

t <- theme_grey()
calc_element('text', t)

# Compare the "raw" element definition to the element with calculated inheritance
t$axis.text.x
calc_element('axis.text.x', t, verbose = TRUE)

# This reports that axis.text.x inherits from axis.text,
# which inherits from text. You can view each of them with:
t$axis.text.x
t$axis.text
t$text

[Package ggplot2 version 3.3.5 Index]