Continuous scale constructor.

Continuous scale constructor.

continuous_scale(aesthetics, scale_name, palette, name = waiver(),
  breaks = waiver(), minor_breaks = waiver(), labels = waiver(),
  limits = NULL, rescaler = rescale, oob = censor,
  expand = waiver(), na.value = NA_real_, trans = "identity",
  guide = "legend", position = "left", super = ScaleContinuous)

Arguments

aesthetics
The names of the aesthetics that this scale works with

scale_name
The name of the scale

palette
A palette function that when called with a numeric vector with values between 0 and 1 returns the corresponding values in the range the scale maps to.

name
The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

breaks
One of:

NULL for no breaks

waiver() for the default breaks computed by the transformation object

A numeric vector of positions

A function that takes the limits as input and returns breaks as output

minor_breaks
One of:

NULL for no minor breaks

waiver() for the default breaks (one minor break between each major break)

A numeric vector of positions

A function that given the limits returns a vector of minor breaks.

labels
One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

A function that takes the breaks as input and returns labels as output

limits
A numeric vector of length two providing limits of the scale. Use NA to refer to the existing minimum or maximum.

rescaler
Used by diverging and n colour gradients (i.e. scale_colour_gradient2(), scale_colour_gradientn()). A function used to scale the input values to the range [0, 1].

oob
Function that handles limits outside of the scale limits (out of bounds). The default replaces out of bounds values with NA.

expand
Vector of range expansion constants used to add some padding around the data, to ensure that they are placed some distance away from the axes. Use the convenience function expand_scale() to generate the values for the expand argument. The defaults are to expand the scale by 5% on each side for continuous variables, and by 0.6 units on each side for discrete variables.

na.value
Missing values will be replaced with this value.

trans
Either the name of a transformation object, or the object itself. Built-in transformations include "asn", "atanh", "boxcox", "exp", "identity", "log", "log10", "log1p", "log2", "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".

A transformation object bundles together a transform, its inverse, and methods for generating breaks and labels. Transformation objects are defined in the scales package, and are called name_trans, e.g. scales::boxcox_trans(). You can create your own transformation with scales::trans_new().

guide
A function used to create a guide or its name. See guides() for more info.

position
The position of the axis. "left" or "right" for vertical scales, "top" or "bottom" for horizontal scales

super
The super class to use for the constructed scale