coord_fixed {ggplot2}	R Documentation

Cartesian coordinates with fixed relationship between x and y scales.

Description

A fixed scale coordinate system forces a specified ratio between the physical representation of data units on the axes. The ratio represents the number of units on the y-axis equivalent to one unit on the x-axis. The default, ratio = 1, ensures that one unit on the x-axis is the same length as one unit on the y-axis. Ratios higher than one make units on the y axis longer than units on the x-axis, and vice versa. This is similar to eqscplot, but it works for all types of graphics.

Usage

  coord_fixed(ratio = 1, xlim = NULL, ylim = NULL,
    wise = NULL)

Arguments

ratio
aspect ratio, expressed as y / x

xlim
limits for the x axis

ylim
limits for the y axis

wise
deprecated in 0.9.1

Examples

# ensures that the ranges of axes are equal to the specified ratio by
# adjusting the plot aspect ratio

qplot(mpg, wt, data = mtcars) + coord_fixed(ratio = 1)
qplot(mpg, wt, data = mtcars) + coord_fixed(ratio = 5)
qplot(mpg, wt, data = mtcars) + coord_fixed(ratio = 1/5)

# Resize the plot to see that the specified aspect ratio is maintained

[Package ggplot2 version 0.9.3.1 Index]