coord_munch {ggplot2}	R Documentation

Munch coordinates data

Description

This function "munches" lines, dividing each line into many small pieces so they can be transformed independently. Used inside geom functions.

Usage

coord_munch(coord, data, range, segment_length = 0.01)

Arguments

coord
Coordinate system definition.

data
Data set to transform - should have variables x and y are chopped up into small pieces (as defined by group). All other variables are duplicated as needed.

range
Panel range specification.

segment_length
Target segment length

[Package ggplot2 version 3.3.5 Index]