facet_null {ggplot2}	R Documentation

Facet specification: a single panel.

Description

Facet specification: a single panel.

Usage

  facet_null(shrink = TRUE)

Arguments

shrink
If TRUE, will shrink scales to fit output of statistics, not raw data. If FALSE, will be range of raw data before statistical summary.

Examples

# facet_null is the default facetting specification if you
# don't override it with facet_grid or facet_wrap
ggplot(mtcars, aes(mpg, wt)) + geom_point()
qplot(mpg, wt, data = mtcars)

[Package ggplot2 version 0.9.3.1 Index]