fortify {ggplot2}	R Documentation

Fortify a model with data.

Description

Rather than using this function, I now recommend using the broom package, which implements a much wider range of methods. fortify() may be deprecated in the future.

Usage

fortify(model, data, ...)

Arguments

model
model or other R object to convert to data frame

data
original dataset, if needed

...
other arguments passed to methods

See Also

fortify.lm()

[Package ggplot2 version 3.4.4 Index]