fortify.map {ggplot2}	R Documentation

Fortify method for map objects.

Description

This function turns a map into a data frame that can more easily be plotted with ggplot2.

Usage

## S3 method for class 'map'
fortify(model, data, ...)

Arguments

model
map object

data
not used by this method

...
not used by this method

See Also

map_data and borders

Examples

if (require("maps")) {
ca <- map("county", "ca", plot = FALSE, fill = TRUE)
head(fortify(ca))
ggplot(ca, aes(long, lat)) +
  geom_polygon(aes(group = group))

tx <- map("county", "texas", plot = FALSE, fill = TRUE)
head(fortify(tx))
ggplot(tx, aes(long, lat)) +
  geom_polygon(aes(group = group), colour = "white")
}

[Package ggplot2 version 2.1.0 Index]