ggplot2-package {ggplot2}	R Documentation

ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics

Description

A system for 'declaratively' creating graphics, based on "The Grammar of Graphics". You provide the data, tell 'ggplot2' how to map variables to aesthetics, what graphical primitives to use, and it takes care of the details.

Author(s)

Maintainer: Thomas Lin Pedersen thomas.pedersen@rstudio.com (ORCID)

Authors:

Hadley Wickham hadley@rstudio.com (ORCID)

Winston Chang (ORCID)

Lionel Henry

Kohske Takahashi

Claus Wilke (ORCID)

Kara Woo (ORCID)

Hiroaki Yutani (ORCID)

Dewey Dunnington (ORCID)

Other contributors:

RStudio [copyright holder, funder]

See Also

Useful links:

https://ggplot2.tidyverse.org

https://github.com/tidyverse/ggplot2

Report bugs at https://github.com/tidyverse/ggplot2/issues

[Package ggplot2 version 3.3.5 Index]