ggplot2-package {ggplot2}	R Documentation

ggplot2: Create Elegant Data Visualisations Using the Grammar of Graphics

Description

A system for 'declaratively' creating graphics, based on "The Grammar of Graphics". You provide the data, tell 'ggplot2' how to map variables to aesthetics, what graphical primitives to use, and it takes care of the details.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Authors:

Winston Chang

Lionel Henry

Thomas Lin Pedersen

Kohske Takahashi

Claus Wilke

Kara Woo

Hiroaki Yutani

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

http://ggplot2.tidyverse.org

https://github.com/tidyverse/ggplot2

Report bugs at https://github.com/tidyverse/ggplot2/issues

[Package ggplot2 version 3.2.1 Index]