ggplot_add {ggplot2}	R Documentation

Add custom objects to ggplot

Description

This generic allows you to add your own methods for adding custom objects to a ggplot with +.gg.

Usage

ggplot_add(object, plot, object_name)

Arguments

object
An object to add to the plot

plot
The ggplot object to add object to

object_name
The name of the object to add

Value

A modified ggplot object

[Package ggplot2 version 3.3.5 Index]