ggplot_gtable {ggplot2}	R Documentation

Build a plot with all the usual bits and pieces.

Description

This function builds all grobs necessary for displaying the plot, and stores them in a special data structure called a gtable(). This object is amenable to programmatic manipulation, should you want to (e.g.) make the legend box 2 cm wide, or combine multiple plots into a single display, preserving aspect ratios across the plots.

Usage

ggplot_gtable(data)

Arguments

data
plot data generated by ggplot_build()

Value

a gtable() object

See Also

print.ggplot() and benchplot() for for functions that contain the complete set of steps for generating a ggplot2 plot.

[Package ggplot2 version 3.3.5 Index]