Save a ggplot (or other grid object) with sensible defaults

ggsave() is a convenient function for saving a plot. It defaults to saving the last plot that you displayed, using the size of the current graphics device. It also guesses the type of graphics device from the extension.

ggsave(filename, plot = last_plot(), device = NULL, path = NULL,
  scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ...)

Arguments

filename

File name to create on disk.

plot

Plot to save, defaults to last plot displayed.

device

Device to use. Can be either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).

path

Path to save plot to (combined with filename).

scale

Multiplicative scaling factor.

width, height, units

Plot size in units ("in", "cm", or "mm"). If not supplied, uses the size of current graphics device.

dpi

Plot resolution. Applies only to raster output types.

limitsize

When TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.

...

Other arguments passed on to graphics device.

Examples

## Not run: ------------------------------------
# ggplot(mtcars, aes(mpg, wt)) + geom_point()
# 
# ggsave("mtcars.pdf")
# ggsave("mtcars.png")
# 
# ggsave("mtcars.pdf", width = 4, height = 4)
# ggsave("mtcars.pdf", width = 20, height = 20, units = "cm")
# 
# unlink("mtcars.pdf")
# unlink("mtcars.png")
# 
# # specify device when saving to a file with unknown extension
# # (for example a server supplied temporary file)
# file <- tempfile()
# ggsave(file, device = "pdf")
# unlink(file)
## ---------------------------------------------