ggscale {ggplot2}	R Documentation

Components of a scale:

Description

Guide related:

Details

name

breaks

labels

expand

Mapping related:

aesthetic

limits

palette

trans

Scales are an S3 class with a single mutable component implemented with a reference class - the range of the data. This mutability makes working with scales much easier, because it makes it possible to distribute the training, without having to worry about collecting all the pieces back together again.

[Package ggplot2 version 1.0.1 Index]