graphical-units {ggplot2}	R Documentation

Graphical units

Description

Multiply size in mm by these constants in order to convert to the units that grid uses internally for lwd and fontsize.

Usage

.pt

.stroke

Format

An object of class numeric of length 1.

An object of class numeric of length 1.

[Package ggplot2 version 3.3.5 Index]