A selection of summary functions from Hmisc

These are wrappers around functions from Hmisc designed to make them easier to use with stat_summary(). See the Hmisc documentation for more details:

Hmisc::smean.cl.boot()

Hmisc::smean.cl.normal()

Hmisc::smean.sdl()

Hmisc::smedian.hilow()

Usage

mean_cl_boot(x, ...)

mean_cl_normal(x, ...)

mean_sdl(x, ...)

median_hilow(x, ...)

Arguments

x

a numeric vector

...

other arguments passed on to the respective Hmisc function.

Value

A data frame with columns y, ymin, and ymax.

Examples

if (requireNamespace("Hmisc", quietly = TRUE)) {
x <- rnorm(100)
mean_cl_boot(x)
mean_cl_normal(x)
mean_sdl(x)
median_hilow(x)
}
#>           y      ymin     ymax
#> 1 0.1338222 -1.649625 1.981843