Generic labeller function for facets

This function makes it easy to assign different labellers to different factors. The labeller can be a function or it can be a named character vectors that will serve as a lookup table.

labeller(..., .rows = NULL, .cols = NULL, keep.as.numeric = NULL,
  .multi_line = TRUE, .default = label_value)

Arguments

...

Named arguments of the form variable = labeller. Each labeller is passed to as_labeller() and can be a lookup table, a function taking and returning character vectors, or simply a labeller function.

.rows, .cols

Labeller for a whole margin (either the rows or the columns). It is passed to as_labeller(). When a margin-wide labeller is set, make sure you don't mention in ... any variable belonging to the margin.

keep.as.numeric

Deprecated. All supplied labellers and on-labeller functions should be able to work with character labels.

.multi_line

Whether to display the labels of multiple factors on separate lines. This is passed to the labeller function.

.default

Default labeller for variables not specified. Also used with lookup tables or non-labeller functions.

Value

A labeller function to supply to facet_grid for the argument labeller.

Details

In case of functions, if the labeller has class labeller, it is directly applied on the data frame of labels. Otherwise, it is applied to the columns of the data frame of labels. The data frame is then processed with the function specified in the .default argument. This is intended to be used with functions taking a character vector such as capitalize.

See also

as_labeller(), labellers

Examples