labs {ggplot2}	R Documentation

Change axis labels and legend titles

Description

Change axis labels and legend titles

Usage

labs(...)

xlab(label)

ylab(label)

ggtitle(label)

Arguments

...
a list of new names in the form aesthetic = "new name"

label
The text for the axis or plot title.

Examples

p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p + labs(title = "New plot title")
p + labs(x = "New x label")
p + xlab("New x label")
p + ylab("New y label")
p + ggtitle("New plot title")

# This should work independently of other functions that modify the
# the scale names
p + ylab("New y label") + ylim(2, 4)
p + ylim(2, 4) + ylab("New y label")

# The labs function also modifies legend labels
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point()
p + labs(colour = "Cylinders")

# Can also pass in a list, if that is more convenient
p + labs(list(title = "Title", x = "X", y = "Y"))

[Package ggplot2 version 2.1.0 Index]