mean_se {ggplot2}	R Documentation

Calculate mean and standard error of the mean

Description

For use with stat_summary()

Usage

mean_se(x, mult = 1)

Arguments

x
numeric vector.

mult
number of multiples of standard error.

Value

A data frame with three columns:

y

The mean.

ymin

The mean minus the multiples of the standard error.

ymax

The mean plus the multiples of the standard error.

Examples

x <- rnorm(100)
mean_se(x)

[Package ggplot2 version 3.3.5 Index]