resolution {ggplot2}	R Documentation

Compute the "resolution" of a numeric vector

Description

The resolution is the smallest non-zero distance between adjacent values. If there is only one unique value, then the resolution is defined to be one. If x is an integer vector, then it is assumed to represent a discrete variable, and the resolution is 1.

Usage

resolution(x, zero = TRUE)

Arguments

x
numeric vector

zero
should a zero value be automatically included in the computation of resolution

Examples

resolution(1:10)
resolution((1:10) - 0.5)
resolution((1:10) - 0.5, FALSE)

# Note the difference between numeric and integer vectors
resolution(c(2, 10, 20, 50))
resolution(c(2L, 10L, 20L, 50L))

[Package ggplot2 version 3.4.4 Index]