scale_colour_brewer {ggplot2}	R Documentation

Sequential, diverging and qualitative colour scales from colorbrewer.org

Description

See http://colorbrewer2.org for more information.

Usage

  scale_colour_brewer(..., type = "seq", palette = 1)

  scale_fill_brewer(..., type = "seq", palette = 1)

  scale_color_brewer(..., type = "seq", palette = 1)

Arguments

type
One of seq (sequential), div (diverging) or qual (qualitative)

palette
If a string, will use that named palette. If a number, will index into the list of palettes of appropriate type

...
Other arguments passed on to discrete_scale to control name, limits, breaks, labels and so forth.

See Also

Other colour scales: scale_color_continuous, scale_color_discrete, scale_color_gradient, scale_color_gradient2, scale_color_gradientn, scale_color_grey, scale_color_hue, scale_colour_continuous, scale_colour_discrete, scale_colour_gradient, scale_colour_gradient2, scale_colour_gradientn, scale_colour_grey, scale_colour_hue, scale_fill_continuous, scale_fill_discrete, scale_fill_gradient, scale_fill_gradient2, scale_fill_gradientn, scale_fill_grey, scale_fill_hue

Examples

dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- qplot(carat, price, data=dsamp, colour=clarity))

# Change scale label
d + scale_colour_brewer()
d + scale_colour_brewer("clarity")
d + scale_colour_brewer(expression(clarity[beta]))

# Select brewer palette to use, see ?scales::brewer_pal for more details
d + scale_colour_brewer(type="seq")
d + scale_colour_brewer(type="seq", palette=3)

d + scale_colour_brewer(palette="Blues")
d + scale_colour_brewer(palette="Set1")

# scale_fill_brewer works just the same as
# scale_colour_brewer but for fill colours
ggplot(diamonds, aes(x=price, fill=cut)) +
  geom_histogram(position="dodge", binwidth=1000) +
  scale_fill_brewer()

[Package ggplot2 version 0.9.3.1 Index]