should_stop {ggplot2}	R Documentation

Used in examples to illustrate when errors should occur.

Description

Used in examples to illustrate when errors should occur.

Usage

should_stop(expr)

Arguments

expr
code to evaluate.

Examples

should_stop(stop("Hi!"))
should_stop(should_stop("Hi!"))

[Package ggplot2 version 3.3.5 Index]