stat_function {ggplot2}	R Documentation

Compute function for each x value

Description

This stat makes it easy to superimpose a function on top of an existing plot. The function is called with a grid of evenly spaced values along the x axis, and the results are drawn (by default) with a line.

Usage

stat_function(mapping = NULL, data = NULL, geom = "path",
  position = "identity", ..., fun, xlim = NULL, n = 101,
  args = list(), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

geom
The geometric object to use display the data

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

fun
Function to use. Either 1) an anonymous function in the base or rlang formula syntax (see rlang::as_function()) or 2) a quoted or character name referencing a function; see examples. Must be vectorised.

xlim
Optionally, restrict the range of the function to this range.

n
Number of points to interpolate along

args
List of additional arguments to pass to fun

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Aesthetics

stat_function() understands the following aesthetics (required aesthetics are in bold):

group

y

Learn more about setting these aesthetics in vignette("ggplot2-specs").

Computed variables

x

x's along a grid

y

value of function evaluated at corresponding x

See Also

rlang::as_function()

Examples

# stat_function is useful for overlaying functions
set.seed(1492)
ggplot(data.frame(x = rnorm(100)), aes(x)) +
  geom_density() +
  stat_function(fun = dnorm, colour = "red")

# To plot functions without data, specify range of x-axis
base <- ggplot(data.frame(x = c(-5, 5)), aes(x))
base + stat_function(fun = dnorm)
base + stat_function(fun = dnorm, args = list(mean = 2, sd = .5))

# The underlying mechanics evaluate the function at discrete points
# and connect the points with lines
base <- ggplot(data.frame(x = c(-5, 5)), aes(x))
base + stat_function(fun = dnorm, geom = "point")
base + stat_function(fun = dnorm, geom = "point", n = 20)
base + stat_function(fun = dnorm, n = 20)

# Two functions on the same plot
base +
  stat_function(fun = dnorm, colour = "red") +
  stat_function(fun = dt, colour = "blue", args = list(df = 1))

# Using a custom anonymous function
base + stat_function(fun = function(.x) .5*exp(-abs(.x)))
base + stat_function(fun = ~ .5*exp(-abs(.x)))

# Using a custom named function
f <- function(.x) .5*exp(-abs(.x))
base + stat_function(fun = f)

[Package ggplot2 version 3.2.1 Index]