tidyeval {ggplot2}	R Documentation

Tidy eval helpers

Description

sym() creates a symbol from a string and syms() creates a list of symbols from a character vector.

expr() and quo() quote one expression. quo() wraps the quoted expression in a quosure.

The plural variants exprs() and quos() return a list of quoted expressions or quosures.

enexpr() and enquo() capture the expression supplied as argument by the user of the current function (enquo() wraps this expression in a quosure).

enexprs() and enquos() capture multiple expressions supplied as arguments, including ....

[Package ggplot2 version 3.2.1 Index]