waiver {ggplot2}	R Documentation

A waiver object.

Description

A waiver is a "flag" object, similar to NULL, that indicates the calling function should just use the default value. It is used in certain functions to distinguish between displaying nothing (NULL) and displaying a default value calculated elsewhere (waiver())

Usage

waiver()

[Package ggplot2 version 3.3.5 Index]